requirements.txt
appdirs==1.4.3
autopep8==1.3.5
awesome-slugify==1.6.5
diff-match-patch==20121119
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.6
django-appconf==1.0.2
django-autofixture==0.12.1
django-ckeditor==5.5.0
django-compressor==2.2
django-extensions==2.0.7
django-import-export==1.0.1
django-js-asset==1.1.0
django-libsass==0.7
django-robots==3.0
ecdsa==0.13
et-xmlfile==1.0.1
flake8==3.5.0
future==0.16.0
get==1.0.3
httplib2==0.11.3
idna==2.6
jdcal==1.4
libsass==0.14.5
markdown2==2.3.5
mccabe==0.6.1
odfpy==1.3.6
olefile==0.45.1
openpyxl==2.5.4
packaging==17.1
Pillow==5.1.0
post==1.0.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
public==1.0.3
pyasn1==0.4.3
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pyflakes==1.6.0
PyGithub==1.40
Pygments==2.2.0
PyJWT==1.6.4
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
python-dateutil==2.7.3
python-instagram==1.3.2
python-jose==3.0.0
pytz==2018.5
PyYAML==3.12
query-string==1.0.2
rcssmin==1.0.6
regex==2018.6.21
request==1.0.2
requests==2.19.1
rjsmin==1.0.12
rsa==3.4.2
ruamel-yaml==0.15.37
simplejson==3.16.0
six==1.11.0
static3==0.7.0
tablib==0.12.1
unicodecsv==0.14.1
Unidecode==0.4.21
urllib3==1.22
xlrd==1.1.0
xlwt==1.3.0
